Move integrations tests to actual folder path
Moved from test/java/org.graylog.integrations to test/java/org/graylog/integrations
