+++ src/test/java/org/graylog/integrations/aws/cloudwatch/CloudWatchServiceTest.java
