+++ src/test/java/org/graylog/integrations/aws/cloudwatch/FlowLogMessageTest.java
@@ -6,7 +6,6 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-
 public class FlowLogMessageTest {
 
     @Test
