+++ src/test/java/org/graylog/integrations/aws/codecs/AWSCodecTest.java
@@ -2,7 +2,7 @@ package org.graylog.integrations.aws.codecs;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.graylog.integrations.AWSTestingUtils;
+import org.graylog.integrations.aws.AWSTestingUtils;
 import org.graylog.integrations.aws.AWSMessageType;
 import org.graylog.integrations.aws.cloudwatch.KinesisLogEntry;
 import org.graylog.integrations.aws.inputs.AWSInput;
