+++ src/test/java/org/graylog/integrations/aws/service/AWSServiceTest.java
