+++ src/test/java/org/graylog/integrations/inputs/paloalto/PaloAltoCodecTest.java
@@ -22,7 +22,6 @@ import org.graylog2.plugin.Message;
 import org.graylog2.plugin.configuration.Configuration;
 import org.graylog2.plugin.journal.RawMessage;
 import org.joda.time.DateTime;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.io.FileWriter;
