+++ src/test/java/org/graylog/integrations/inputs/paloalto/PaloAltoTemplateTest.java
